submissions/sachin_2024-07-02/document_text_extractor_app/requirements.txt
fastapi
streamlit
pymupdf
uvicorn
python-dotenv
